autonomus-hr-chatbot/requirements.txt
langchain
streamlit
streamlit-chat
pinecone-client
pandas
python-dotenv
